Bringing it all in for EB
